# docs/api.yaml
openapi: 3.0.3
info:
  title: Flask MongoDB API
  description: REST API for managing customers and subscriptions
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/products:
    get:
      summary: Get all products
      description: Retrieve a list of products with optional filtering and pagination
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: is_active
          in: query
          description: Filter by product active status
          schema:
            type: boolean
        - name: billing_cycle
          in: query
          description: Filter by billing cycle
          schema:
            type: string
            enum: [weekly, monthly, yearly]
        - name: customizable
          in: query
          description: Filter by customizable products
          schema:
            type: boolean
        - name: search
          in: query
          description: Search products by name or description
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a new product (AddProduct)
      description: Create a new product with the provided data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Product with name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/products/{product_id}:
    get:
      summary: Get a product by ID
      description: Retrieve a specific product by its ID
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a product
      description: Update an existing product's information
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error or invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a product
      description: Delete a product (only if it has no active subscriptions)
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Product deleted successfully
        "400":
          description: Invalid ID or product has active subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/products/{product_id}/activate:
    post:
      summary: Activate a product
      description: Set product status to active
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Product activated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/products/{product_id}/deactivate:
    post:
      summary: Deactivate a product
      description: Set product status to inactive
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Product deactivated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 10
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Resource not found"
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: object
          example: { "email": ["Invalid email format"] }
    Product:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Premium Plan"
        description:
          type: string
          example: "Premium subscription plan with advanced features"
        price:
          type: number
          example: 29.99
        currency:
          type: string
          enum: [USD, EUR, GBP, JPY]
          example: "USD"
        billing_cycle:
          type: string
          enum: [weekly, monthly, yearly]
          example: "monthly"
        trial_period_days:
          type: integer
          minimum: 0
          maximum: 365
          example: 14
        features:
          type: array
          items:
            type: string
          example:
            [
              "Priority Support",
              "Unlimited Storage",
              "Advanced Analytics",
              "API Access",
            ]
        is_active:
          type: boolean
          example: true
        customizable:
          type: boolean
          example: true
          description: "Whether the product is customizable"
        customizable_fields:
          type: array
          items:
            type: string
          example:
            [
              "topBarColor",
              "topBarBackgroundColor",
              "topBarButtonBackgroundColor",
              "positionIndex",
              "defaultLang",
            ]
          description: "Array of customizable field names"
        default_settings:
          type: object
          example:
            {
              "topBarColor": "#ffffff",
              "topBarBackgroundColor": "#000000",
              "topBarButtonBackgroundColor": "#000000",
              "positionIndex": 0,
              "defaultLang": "en",
            }
          description: "Default settings for customizable products"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductCreate:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Premium Plan"
        description:
          type: string
          maxLength: 200
          example: "Premium subscription plan with advanced features"
        price:
          type: number
          minimum: 0
          example: 29.99
        currency:
          type: string
          enum: [USD, EUR, GBP, JPY]
          default: USD
        billing_cycle:
          type: string
          enum: [weekly, monthly, yearly]
          default: monthly
        trial_period_days:
          type: integer
          minimum: 0
          maximum: 365
          example: 14
        features:
          type: array
          items:
            type: string
          example:
            [
              "Priority Support",
              "Unlimited Storage",
              "Advanced Analytics",
              "API Access",
            ]
        is_active:
          type: boolean
          example: true
        customizable:
          type: boolean
          example: true
          description: "Whether the product is customizable"
        customizable_fields:
          type: array
          items:
            type: string
          example:
            [
              "topBarColor",
              "topBarBackgroundColor",
              "topBarButtonBackgroundColor",
              "positionIndex",
              "defaultLang",
            ]
          description: "Array of customizable field names"
        default_settings:
          type: object
          example:
            {
              "topBarColor": "#ffffff",
              "topBarBackgroundColor": "#000000",
              "topBarButtonBackgroundColor": "#000000",
              "positionIndex": 0,
              "defaultLang": "en",
            }
          description: "Default settings for customizable products"
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Premium Plan"
        description:
          type: string
          maxLength: 200
          example: "Premium subscription plan with advanced features"
        price:
          type: number
          minimum: 0
          example: 29.99
        currency:
          type: string
          enum: [USD, EUR, GBP, JPY]
        billing_cycle:
          type: string
          enum: [weekly, monthly, yearly]
        trial_period_days:
          type: integer
          minimum: 0
          maximum: 365
        features:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        customizable:
          type: boolean
        customizable_fields:
          type: array
          items:
            type: string
        default_settings:
          type: object
...